requirements.txt
# requirements.txt

# Framework central
pytest>=8.0.0,<9.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-xdist>=3.0.0,<4.0.0
trcli>=1.9.13

# Integração TestRail 
pytest-testrail>=2.9.0
requests>=2.28.0,<3.0.0

# Utilitários
trimesh>=4.6.11
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0
mypy>=1.0.0,<2.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-html>=3.1.0,<4.0.0
pytest-json-report>=1.5.0,<2.0.0
packaging>=21.0
pluggy>=1.0.0,<2.0.0
mutmut

# API
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
httpx
python-multipart
python-dotenv

# Dependências DB
SQLAlchemy>=2.0.0,<3.0.0
mysql-connector-python>=8.0.0,<9.0.0
alembic
